Remove translations
This is done for the following reasons..

Browsers such as Firefox and Chromium are able to translate
on the fly nowadays. The quality of these translations is getting
better and better.

Although Launchpad translators provide better/exact translations they
don't do so in real time. Whenever a change occurs in the git tree,
the affected chapter goes back to English. This causes two issues:

- The document is in two languages: Local language and English.
- We're hesitant to modify the document for this not to happen, which
causes bad maintenance. We don't rephrase until absolutely necessary,
we don't update information unless it's critically important.

The biggest issue now: the syntax used is complex.
Many translators break it.

In Python or C, we're using %s or %d markers which translators rarely
break. When it happens we detect it also thanks to internal tools.

In documents, the syntax relies on a series of quotes and inverted
quotes which translators often replace... a " is turned into a «, a ``
is turned into “..etc.

Throughout the years this led to the installation guide only being
translated fully in a few languages and new translations for it being
ignored because they introduce breakages.

The last time the translations were updated was in 2018. It's been
in the TODO list for 6 years now!

Last but not least, each translation requires its git hook to trigger
an RTFM build. Github is limited in number of hooks so this forced
us to host this project on Gitlab.